Add hook for defining absolute library path
While the code to locate libraries is useful it will step on the toes of
packagers which need hard control of load path. Patch adds a system
property harc.decodeir.library.path which if set is used as an absolute
path to library. If unset, the old behaviour is unchanged.