rxtx: Add a generic rxtxcomm device
The source contains a "Linux-alltargets" configuration target which
contains ttyACM and some other less used devices. It could be argued
that this could be used nowadays, since the original reason not to use
it was the very large number of /dev/* nodes before udev reduced this
to something more sane. However, the status of this code is generally
questionable, at least some device names are probably out of date.

Instead this patch also adds a /dev/rxtxcomm* device. There is no driver
using this, but it can be used as a udev link for any device not
otherwise supported.